requirements.txt
config-parser
paho-mqtt